First pass of initialization refactoring.
To be able to expose the `pg_autoctl create ...` operation at the HTTP API
level, we need to move some of the processing to the normal FSM transitions.
This patch moves the `pg_ctl initdb` operation to the transition function
from INIT to SINGLE, when this happens during the init step.