Some progress towards /api/1.0/fsm/assign.
We use file name pattern matching (glob) for the URLs now, which allow us to
match the last part of the URI with any goal state to assign. That's nice.
There's still some problem in the Keeper structure initialisation that
prevents us from actually doing the work, that's for another day now.