Improve the listener process and PIPE communication channel.
Now when the HTTPd server needs to execute a sub-command, it does so by
sending the command to the listener process. This process is responsible for
running the sub-command and getting the result back to the httpd process.

That way, the subcommands don't run in a Unix process environment where the
HTTP port we're using is open, and without inheritting other aspects of our
pg_autoctl services. That's cleaner.

For this to work we also have to improve the starting/stoping and signal
handling of the different processes that are all together implementing the
pg_autoctl service, and a first version of that code is included in this
patch.